// src/types/PlayerLoot.ts
export type PlayerLoot = {
  asset?: string;
  count: number;
  disenchantLootName?: string;
  disenchantValue?: number;
  displayCategories?: string;
  expiryTime?: number;
  isNew?: boolean;
  isRental?: boolean;
  itemDesc: string;
  itemStatus?: string;
  localizedDescription?: string;
  localizedName: string;
  localizedRecipeSubtitle?: string;
  localizedRecipeTitle?: string;
  lootId: string;
  // Original property meant for storing untranslated loot name
  lootNameRaw?: string;
  lootName: string;
  parentItemStatus?: string;
  parentStoreItemId?: number;
  rarity: string;
  redeemableStatus?: string;
  refId?: string;
  rentalGames?: number;
  rentalSeconds?: number;
  shadowPath?: string;
  splashPath?: string;
  storeItemId?: number;
  tags?: string;
  tilePath: string;
  type: string;
  upgradeEssenceName?: string;
  upgradeEssenceValue?: number;
  upgradeLootName?: string;
  value?: number;
};